Allow the ACPI handle for a PCI device to be overridden instead of using the corresponding handle in the device tree.
At one point this seemed to partially help with getting acpi_video to
work on my ThinkPad x220.  However, the real solution to that seems to
have been enabling opregion support in the i915kms driver instead.
Given that, I don't know of a use case for this, but I'm parking it here
in case a future one turns up.