Fix early boot.
Now that the FPU save area is variable sized after the pcb, the
true top of stack for thread0 can't be calculated until npxinit()
runs.  Only use the bottom page of proc0's stack for init386() and
have init386() return the true top of stack to locore which then
adjusts it before calling mi_startup().