Rework buffer management to support better PUSH handling.
This still uses the PUSH timer for now (so each PUSH packet is sent
to a separate DDP buffer).  However, these changes should allow us to
switch to using PUSH-no-invalidate instead.  We now track a read_index
into the buffer as well as update the write index of the static DDP
buffers.  On each read, userland is given the currently unread chunk
of the buffer and the read index is advanced.  In addition,
handle_ddp_data() will now accept a DDP report that doesn't invalidate
the buffer but reports the PSH flag instead (but only for a static
DDP connection).