Import the inspect tool and INSPECTION utility.
This is a hacky tool to log messages on the console if certain operations
take a "long" time.  Right now you can set thresholds to log if a lock is
held for too long, if a thread is preempted for too long, if a thread
blocks or spins on a lock for too long, or if a thread sleeps for too long.
The noise is trimmed by tagging specific processes to monitor rather than
monitoring all threads in the system.

Right now this uses a couple of sysctls to get and set the inspection flags
because this initially predated procctl().  I might change this to use
procctl() instead.