Make /dev/tap and /dev/vmnet static devs and use cdevpriv.
Tap file descriptors now store a reference to the tap in the cdevpriv.
For /dev/tapX devices, the reference is obtained from si_drv1.  For
the cloning devices, a free interface is located or created and stored
in cdevpriv.  The cdevpriv holds a reference on the cdev which is
released when tap_dtor runs.