Add kvm_nlist2() and struct kvm_nlist.
The nlist structure uses u_long for n_value, but that would result in
address truncation when a 32-bit target inspects a 64-bit core.  Instead,
introduce a kvm_nlist structure which uses kvadd_r for n_value and use
this internally.  kvm_nlist() still returns a struct nlist and is only
permitted for native cores.