Change the pru_aio interface to be active.
Now pru_aio_queue() can "claim" a request by returning 0.  In this
case it is responsible for maintaining its own internal queue of
requests.  Only if pru_aio_queue() claims a request is pru_aio_cancel()
called to cancel a request.  It can fail to cancel a request by
returning a non-zero value.