Don't wait for more data after a short read.
AIO reads imply non-blocking I/O.  If DDP is not enabled, then we would
copy the pending data and wait for more.  Now we always complete a
request if we copy out any data from the socket.  (This should probably
be honoring 'lowat' though.)

For DDP, update the comments (but #ifdef the code) to always flush after
doing a copy just to give the NIC a chance to place any currently queued
data to go along with the indicate.  I'm not sure if this is strictly
needed (e.g. is a PSH from an indicate "sticky" and force a completion
of the next DDP?), hence the #ifdef notyet.