Fix initial buffer offset.
This is not the page offset of the start of the buffer (that is stored in
the page pod).  Instead, this indicates how much of the data has already
been used.  This should work in conjunction with the push timer to remove
the need for an explicit flush for partially filled requests.  Previously
I believe this was always non-zero resulting in all DDP requests always
completeing.