Only use nm_i for the rx buffers.
Previously the driver tried to use netmap indices for both the
completion ring and the receive buffer ring.  However, the completion
ring is twice as large as the buffer ring by default.  Instead, use
the normal driver method to walk the completion ring (vxcr_next) and
only expose the receive buffers in the buffer ring to netmap.

This also means that we fully handle any completion ring updates during
the first half of RX sync (including synching the head index if
requested).

The second half of RX sync only writes to the receive buffer ring to
queue netmap buffers.