Preserve raw errno value and provide ABI-specific strerror.
For native ABIs this doesn't change anything, but for non-native ABIs
the errno value output will now be the actual value returned to the
application, but the ABI hook can provide a suitable description.  For
now the non-native ABIs convert the native errno to a FreeBSD errno
and use strerror().