Checkpoint WIP.
This expands the aio "library" but also simplifies some things.  Instead
of having an aio_queue(), each request whose queue routine doesn't fail
is implicitly queued.  However, a routine is supplied to install a cancel
routine, and it can chaned as a request changes state.  The caller is
required to complete/cancel the request if the cancel state can't be
changed.

It also starts on an alternate backend for sockets.  I'm going to take
this into a different direction, but I will keep at least some of this.
Requests are now queued on the socket buffer rather than the socket,
and the queueing is self-contained.  This version uses its own taskqueue
where the task is per-sockbuf (rather than per-request) and uses
vm_fault_hold_quick_pages()) to hold the AIO.  The idea is that it
would do a non-blocking soreceive() that returned the raw mbufs and
would then use uiomove_fromphys() to copy the mbufs out.