Rework the TDB_EXIT reporting to avoid sleeping with the tidhash lock held.
To avoid losing the race of the last thread exiting, add a count of "exiting"
threads.  A thread marks itself as exiting while it is blocked in ptracestop()
as well as while the PROC_LOCK is dropped to remove the thread from the
tidhash.  Each thread compares p_exitingthreads against p_numthreads to
determine if it is the last thread.  If it is, it waits for any "exiting"
threads to finish exiting before it returns.
Instead, add a count of "exiting" threads